requirements.txt
Flask==3.0.0
pymysql==1.1.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==8.0.0